+++ dalesbred/src/main/java/org/dalesbred/dialect/Dialect.java
@@ -26,7 +26,7 @@ import org.dalesbred.DatabaseException;
 import org.dalesbred.DatabaseSQLException;
 import org.dalesbred.connection.ConnectionProvider;
 import org.dalesbred.connection.DataSourceConnectionProvider;
-import org.dalesbred.conversion.TypeConversion;
+import org.dalesbred.internal.instantiation.TypeConversion;
 import org.dalesbred.conversion.TypeConversionRegistry;
 import org.dalesbred.internal.jdbc.ArgumentBinder;
 import org.dalesbred.transaction.TransactionManager;