+++ dalesbred/src/main/java/org/dalesbred/dialect/PostgreSQLDialect.java
@@ -51,6 +51,6 @@ public class PostgreSQLDialect extends Dialect {
 
     @Override
     public void registerTypeConversions(@NotNull TypeConversionRegistry typeConversionRegistry) {
-        typeConversionRegistry.registerNonNullConversionToDatabaseType(Date.class, Timestamp.class, v -> new Timestamp(v.getTime()));
+        typeConversionRegistry.registerConversionToDatabase(Date.class, Timestamp.class, v -> new Timestamp(v.getTime()));
     }
 }