+++ dalesbred/src/main/java/org/dalesbred/internal/instantiation/ConversionMap.java
@@ -22,7 +22,6 @@
 
 package org.dalesbred.internal.instantiation;
 
-import org.dalesbred.conversion.TypeConversion;
 import org.jetbrains.annotations.NotNull;
 
 import java.lang.reflect.Type;