+++ dalesbred/src/main/java/org/dalesbred/internal/instantiation/ReflectionInstantiator.java
@@ -22,7 +22,6 @@
 
 package org.dalesbred.internal.instantiation;
 
-import org.dalesbred.conversion.TypeConversion;
 import org.dalesbred.internal.utils.Throwables;
 import org.jetbrains.annotations.NotNull;